move filter to results directory
