--- src/Components/UrgentNeedsTable/UrgentNeedsRow.css
@@ -1,4 +1,0 @@
-button.urgent-need-button {
-  color: white;
-  margin-top: 1rem;
-}
