--- src/Components/UrgentNeedsTable/UrgentNeedsTable.css
@@ -1,6 +1,0 @@
-.urgentNeedsTableFooter {
-  text-align: center;
-  line-height: 1.25rem;
-  font-size: 14px;
-  color: #2a2b2a;
-}
